backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg2-binary>=2.9.10
pgvector>=0.3.6
pydantic[email]>=2.11.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
langchain>=0.3.13
langchain-google-genai>=2.0.7
langchain-community>=0.3.13
pypdf>=5.1.0
httpx>=0.28.1
python-dotenv>=1.0.1
cloudinary>=1.41.0
bcrypt>=4.0.0

# Testing
pytest>=8.0.0
